stg_prediction/requirements.txt
dtaidistance==2.3.10
geopandas=0.6.1=py_0
numpy==1.21.5
tqdm==4.64.0
pytorch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
scipy==1.7.3
matplotlib==3.5.2